requirements.txt
google-api-core
google-cloud-compute
google-cloud-dns
google-api-python-client